+++ market/market/apps/social/admin.py
@@ -1,6 +1,8 @@
 from django.contrib import admin
 
-from market.apps.social.models import SocialProfile
+from market.apps.social.models import (Review,
+                                       SocialProfile)
 
 
+admin.site.register(Review)
 admin.site.register(SocialProfile)
