+++ market/market/apps/social/models.py
@@ -1,4 +1,6 @@
 from django.db import models
+from django_extensions.db.models import (ActivatorModel,
+                                         TimeStampedModel)
 from django.urls import reverse
 
 from djgeojson.fields import PointField
@@ -6,6 +8,21 @@ from djgeojson.fields import PointField
 from market.apps.core.models import (RandomSlugModel,
                                      UserProfile)
 
+class Review(RandomSlugModel, TimeStampedModel):
+    reviewer = models.ForeignKey(UserProfile, on_delete=models.CASCADE, related_name='reviewer')
+    reviewee = models.ForeignKey(UserProfile, default=1, on_delete=models.CASCADE, related_name='reviewee')
+    title = models.CharField(max_length=128)
+    SCORE_CHOICES = [
+                    (1, '1'),
+                    (2, '2'),
+                    (3, '3'),
+                    (4, '4'),
+                    (5, '5'),
+        ]
+    score = models.IntegerField(choices=SCORE_CHOICES, default=4)
+    body = models.TextField(max_length=5000)
+    def get_absolute_url(self):
+        return reverse('messaging:review_deteail', kwargs={'slug': self.slug})
 
 class SocialProfile(RandomSlugModel):
     owner = models.OneToOneField(UserProfile, on_delete=models.CASCADE, related_name='social')
