+++ market/market/apps/social/urls.py
@@ -1,6 +1,9 @@
 from django.conf.urls import url
 
-from market.apps.social.views import (SocialProfileDetailView,
+from market.apps.social.views import (ReviewCreateView,
+                                      ReviewDetailView,
+                                      ReviewListView,
+                                      SocialProfileDetailView,
                                       SocialProfileSelfDetailView,
                                       # SocialProfileListView,
                                       SocialProfileUpdateView)
@@ -14,4 +17,9 @@ urlpatterns = [
 
     # Profile is updated in settings
     url(r'^settings/profile/$', SocialProfileUpdateView.as_view(), name='update'),
+
+    # Review urls
+    url(r'^reviews/new/(?P<slug>[-\w]+)/$', ReviewCreateView.as_view(), name='new_review'),
+    url(r'^reviews/(?P<slug>[-\w]+)/$', ReviewDetailView.as_view(), name='detail_review'),
+    url(r'profile/(?P<slug>[-\w]+)/reviews^$', ReviewListView.as_view(), name='reviews_for_user'),
 ]
